scripts/requirements.txt
requests==2.28.2
Pillow==9.4.0
